feat: Refactor event processing by reorganizing code and adding new event header files
